removing unused queue code
