+++ .flaskenv
@@ -12,8 +12,5 @@ AWS_SECRET_ACCESS_KEY=fsdlrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY # pragma: allowli
 AWS_BUCKET_NAME=fsd-bucket
 AWS_ENDPOINT_OVERRIDE=http://localhost:4566
 AWS_REGION=eu-west-2
-AWS_DLQ_MAX_RECIEVE_COUNT=3
-AWS_SQS_IMPORT_APP_PRIMARY_QUEUE_URL=http://localhost:4566/000000000000/import-queue.fifo
-AWS_SQS_IMPORT_APP_SECONDARY_QUEUE_URL=http://localhost:4566/000000000000/import-queue-dlq.fifo
 
 SQLALCHEMY_WARN_20=1
