+++ application_store/api/__init__.py
@@ -1,2 +1,1 @@
 from application_store.api.routes.application.routes import ApplicationsView  # noqa
-from application_store.api.routes.queues.routes import QueueView  # noqa
