+++ assessment_store/db/queries/__init__.py
@@ -1,5 +1,4 @@
 from .assessment_records.queries import (
-    bulk_insert_application_record,
     delete_assessment_record,
     find_answer_by_key_runner,
     get_metadata_for_fund_round_id,
@@ -18,7 +17,6 @@ __all__ = [
     "select_tags_for_fund_round",
     "insert_tags",
     "get_metadata_for_fund_round_id",
-    "bulk_insert_application_record",
     "find_answer_by_key_runner",
     "get_scores_for_app_sub_crit",
     "create_score_for_app_sub_crit",
