--- copilot/fsd-pre-award-stores/addons/assessment-import-queue.yml
@@ -1,31 +1,0 @@
-Parameters:
-  App:
-    Type: String
-    Description: Your application's name.
-  Env:
-    Type: String
-    Description: The environment name your service, job, or workflow is being deployed to.
-  Name:
-    Type: String
-    Description: The name of the service, job, or workflow being deployed.
-
-Resources:
-  AssessmentImportQueuePolicy:
-    Type: AWS::IAM::ManagedPolicy
-    Properties:
-      PolicyDocument:
-        Version: "2012-10-17"
-        Statement:
-          - Sid: AssessmentImportQueuePolicy
-            Effect: Allow
-            Action:
-              - sqs:SendMessage
-              - sqs:ReceiveMessage
-              - sqs:DeleteMessage
-            Resource:
-              - Fn::ImportValue: !Sub ${App}-${Env}-AssessmentImportQueueArn
-
-Outputs:
-  AssessmentImportQueuePolicyArn:
-    Description: "The ARN of the ManagedPolicy to attach to the task role."
-    Value: !Ref AssessmentImportQueuePolicy
