+++ tests/assessment_store_tests/_helpers.py
@@ -6,9 +6,9 @@ from sqlalchemy.inspection import inspect
 from sqlalchemy.orm import defer
 
 from assessment_store.db.models.assessment_record import AssessmentRecord
-from assessment_store.db.queries import bulk_insert_application_record
 from db import db
 from tests.assessment_store_tests._db_seed_data import get_dynamic_rows
+from tests.assessment_store_tests.conftest import bulk_insert_application_record
 
 
 @contextmanager
